libs/src/change-notifier/pyproject.toml
[tool.poetry]
name = "change-notifier"
version = "0.1.0"
description = ""
authors = ["Antonio Cerciello <antonio.cerciello@gmail.com>"]
readme = "README.md"
packages = [{include = "change_notifier"}]

[tool.poetry.dependencies]
python = "^3.10"
memphis-py = "^1.3.1"
pydantic = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
